Moved scanner interface to engine package
