+++ review/pkg/engine/scanner.go
@@ -1,6 +1,6 @@
-package spider
+package engine
 
 // Позволяет сканировать сайты
-type Interface interface {
+type Scanner interface {
 	Scan(string, int) (map[string]string, error)
 }
